Еще одно исправление, кривого места.

+++ Workwear/Domain/Regulations/NormItem.cs
@@ -131,7 +131,7 @@ namespace workwear.Domain.Regulations
 			double months = amount * oneItemByMonths;
 			int wholeMonths = (int)months;
 			int addintionDays = (int)Math.Round((months - wholeMonths) * 30);
-			if(Math.Abs(wholeMonths) > 100000) {
+			if(Math.Abs((long)wholeMonths) > 100000) {
 				logger.Warn("Расчет периода вышел за 100000 месяцев. Скорей всего изначальные данные некорректны.");
 				return null;
 			}
